updated with a better algorithm
tried a better, holistic way of solving the problem of calculating the probability of getting an offspring with a dominant phenotype if two randomly selected individuals from a population are mated.  the newer method can be used to calculate other stats, such as the probability of getting a recessive offspring, etc.